made heading data type (int vs double) consistent to double

+++ src/slogo/controller/Controller.java
@@ -130,7 +130,7 @@ public class Controller {
      * @param startingY the y position of where the turtle will start
      * @param startingHeading where the turtle will be facing
      */
-    public void addTurtle(String name, double startingX, double startingY, int startingHeading){
+    public void addTurtle(String name, double startingX, double startingY, double startingHeading){
         Turtle t = new Turtle(name, startingX, startingY, startingHeading);
         if(nameCount.containsKey(t.getName())){
             throw new InvalidTurtleException("Turtle already exists", new Throwable()); //shouldn't ever get to this
